refac: changing the typing of production,mortalies and proportion to BigDecimal

+++ src/main/java/br/com/aviapp/api/application/dto/AviaryReportDTO.java
@@ -1,10 +1,11 @@
 package br.com.aviapp.api.application.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record AviaryReportDTO(String name, List<CollectEggDataDTO> eggCollects, List<CollectChickenDTO> chickenCollect,
                               int totalEggsCollected, int totalDeadBirds, int totalDeadChickens, int totalDeadRoosters,
-                              int currentChickens, int currentRoosters, int totalBirds, float production,
-                              float roosterMortality, float chickenMortality, float mortality,
-                              float chickenRoosterProportion, List<EggDetailDTO> quantityByEggType) {
+                              int currentChickens, int currentRoosters, int totalBirds, BigDecimal production,
+                              BigDecimal roosterMortality, BigDecimal chickenMortality, BigDecimal mortality,
+                              BigDecimal chickenRoosterProportion, List<EggDetailDTO> quantityByEggType) {
 }
